Do not test value-based slicing in generic slicing test file
